requirements.txt
alpaca-py
websockets
sqlalchemy
python-dotenv
pytz
